+++ apps/commerce-api/src/main/java/com/loopers/domain/point/PointRepository.java
@@ -1,0 +1,8 @@
+package com.loopers.domain.point;
+
+import java.util.Optional;
+
+public interface PointRepository {
+
+    Optional<Point> findByUserId(String userId);
+}
