+++ apps/commerce-api/src/main/java/com/loopers/domain/point/PointUseCase.java
@@ -1,4 +1,4 @@
-package com.loopers.application.point.port.in;
+package com.loopers.domain.point;
 
 public interface PointUseCase {
