+++ apps/commerce-api/src/main/java/com/loopers/domain/user/UserRegisterCommand.java
@@ -1,4 +1,4 @@
-package com.loopers.application.user.port.in;
+package com.loopers.domain.user;
 
 import com.loopers.domain.user.vo.Gender;
 import lombok.AccessLevel;
