+++ apps/commerce-api/src/main/java/com/loopers/domain/user/UserUseCase.java
@@ -1,4 +1,4 @@
-package com.loopers.application.user.port.in;
+package com.loopers.domain.user;
 
 public interface UserUseCase {
