+++ apps/commerce-api/src/main/java/com/loopers/infrastructure/point/PointJpaRepository.java
@@ -1,6 +1,5 @@
-package com.loopers.infrastructure.point.jpa;
+package com.loopers.infrastructure.point;
 
-import com.loopers.infrastructure.point.entity.PointEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
