+++ apps/commerce-api/src/test/java/com/loopers/application/likes/ProductLikeServiceTest.java
@@ -1,6 +1,7 @@
 package com.loopers.application.likes;
 
 import com.loopers.domain.likes.ProductLikeRepository;
+import com.loopers.domain.likes.ProductLikeService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
