+++ apps/commerce-api/src/test/java/com/loopers/application/product/ProductServiceTest.java
@@ -3,6 +3,8 @@ package com.loopers.application.product;
 import com.loopers.domain.likes.ProductLikeRepository;
 import com.loopers.domain.product.Product;
 import com.loopers.domain.product.ProductRepository;
+import com.loopers.domain.product.ProductResult;
+import com.loopers.domain.product.ProductService;
 import com.loopers.domain.product.brand.Brand;
 import com.loopers.domain.product.brand.BrandRepository;
 import org.junit.jupiter.api.DisplayName;
