+++ apps/commerce-api/src/test/java/com/loopers/application/user/UserServiceTest.java
@@ -1,9 +1,10 @@
 package com.loopers.application.user;
 
-import com.loopers.application.user.exception.UserException;
-import com.loopers.application.user.port.in.UserInfoResult;
-import com.loopers.application.user.port.in.UserRegisterCommand;
+import com.loopers.domain.user.exception.UserException;
+import com.loopers.domain.user.UserInfoResult;
+import com.loopers.domain.user.UserRegisterCommand;
 import com.loopers.domain.user.User;
+import com.loopers.domain.user.UserService;
 import com.loopers.domain.user.fixture.UserFixture;
 import com.loopers.domain.user.fixture.UserRegisterCommandFixture;
 import com.loopers.infrastructure.user.UserRepositoryImpl;
